+++ lib/mvc_rocket.dart
@@ -18,3 +18,4 @@ export 'src/mc_value_listenable.dart';
 export 'src/mc_mini_view.dart';
 export 'src/mc_exception.dart';
 export 'src/mc_constants.dart';
+export 'src/mc_llistenable.dart';